The user can now navigate to a blank drink page
This can be done through pressing one of the drinks on the front page.
