+++ components/Drink.js
@@ -1,0 +1,11 @@
+import {View} from "react-native";
+
+const Drink = () => {
+    return (
+        <View>
+
+        </View>
+    )
+}
+
+export default Drink;
