+++ components/DrinkList.js
@@ -1,14 +1,14 @@
-import {StyleSheet, FlatList, Text, View, Image} from "react-native";
+import {StyleSheet, FlatList, Text, View, Image, TouchableOpacity} from "react-native";
 import drinks from "./DrinkList.TestData.js";
 
-const DrinkList = () => {
+const DrinkList = ({ navigation }) => {
     return (
         <View style={styles.container}>
             <FlatList
                 data={drinks}
                 renderItem={
                     ({item}) => (
-                        <View style={styles.item}>
+                        <TouchableOpacity style={styles.item} onPress={() => navigation.navigate("Drink")}>
                             <View style={styles.pictureContainer}>
                                 <Image source={item.picture} style={styles.picture}></Image>
                             </View>
@@ -16,7 +16,7 @@ const DrinkList = () => {
                                 <Text style={styles.name}>{item.name}</Text>
                                 <Text numberOfLines={2} style={styles.description}>{item.description}</Text>
                             </View>
-                        </View>
+                        </TouchableOpacity>
                     )
                 }
             />
